infra/terraform/scripts/initialise-backend/requirements.txt
boto3
python-hcl2
